Update during Shaftburied
All changes below are during a currently running fortress. YMMV.
A few of the changes have not been directly tested but are due to things
noticed during the running of the fortress.

Changes to the sungwood tree definitions are thanks to 'plant', and I
look forward to always being able to have a sungwood tree on every map
thanks to it. Especially for testing.

I've blindly stumbled onto checking hist fig lists... so a rework of
conscript is likely.

I put in comments to my onLoad.init file on the of chance anyone looks
at it and wonders... why x y or z.

The sungwood reaction file -requiring the S.L.A.D.E Mixery- is a bit of a
long time coming. Never noticed the lack of Sungwood Wine until now.
Mostly I've been fighting the fact that DF expects all plant_growth to be
bag items... so getting them to display if it is a flower or a leaf has
eaten hours of my life.

COOOOOOOMMMMMIIIIT!
 Changes to be committed:
	modified:   data.43.05/init/embark_profiles.txt
	modified:   data.43.05/save/onLoad.init
	modified:   raw-notes.43.05/dfhack.init
	modified:   raw-notes.43.05/onLoad.init
	modified:   raw-notes.43.05/scraps.console.notes
	modified:   raw-notes.43.05/scraps.lua.notes
	modified:   raw-notes.43.05/scraps.rb.notes
	new file:   raw-notes.43.05/sungwood reactions-finally.nfo
	modified:   raw.43.05/objects/creature_domestic.txt
	modified:   raw.43.05/objects/creature_faberge_chicken.txt
	modified:   raw.43.05/objects/entity_angel.txt
	modified:   raw.43.05/objects/entity_animal_peoples.txt
	modified:   raw.43.05/objects/entity_cultist.txt
	modified:   raw.43.05/objects/entity_evil.txt
	modified:   raw.43.05/objects/entity_fab.txt
	modified:   raw.43.05/objects/entity_fallen.txt
	modified:   raw.43.05/objects/entity_forest.txt
	modified:   raw.43.05/objects/entity_gardohn.txt
	modified:   raw.43.05/objects/entity_kapa.txt
	modified:   raw.43.05/objects/entity_kolcha.txt
	modified:   raw.43.05/objects/entity_loli.txt
	modified:   raw.43.05/objects/entity_mewli.txt
	modified:   raw.43.05/objects/entity_mla.txt
	modified:   raw.43.05/objects/entity_mountain.txt
	modified:   raw.43.05/objects/entity_pedo.txt
	modified:   raw.43.05/objects/entity_plains.txt
	modified:   raw.43.05/objects/entity_proudhorn.txt
	modified:   raw.43.05/objects/entity_skulking.txt
	modified:   raw.43.05/objects/entity_vilehorn.txt
	modified:   raw.43.05/objects/entity_wild_foocubi.txt
	modified:   raw.43.05/objects/entity_wraitha.txt
	modified:   raw.43.05/objects/entity_wraithb.txt
	modified:   raw.43.05/objects/entity_wraithc.txt
	modified:   raw.43.05/objects/entity_wraithd.txt
	modified:   raw.43.05/objects/entity_yiffian.txt
	new file:   raw.43.05/objects/onLoad.init
	modified:   raw.43.05/objects/plant_sungwood.txt
	new file:   raw.43.05/objects/reaction_sungwood.txt