44.09 initial upload part 2
And now I'e done the Ulraedit shuiffle to the new directory. I hate this
step. So of course I forgot it a minute ago.

the lazy fab reactions are still in there -now for several versions- as
I'm pondering yet another new material. Or maybe a repurpose. Dunno.

Hence 'pondering'.
 Changes to be committed:
	new file:   raw-notes.44.09/entity_default.44.09.nfo
	new file:   raw-notes.44.09/scraps.console.notes
	new file:   raw-notes.44.09/scraps.console.notes.lazy_fab_reactions
	new file:   raw-notes.44.09/scraps.lua.common
	new file:   raw-notes.44.09/scraps.lua.notes
	new file:   raw-notes.44.09/scraps.rb.notes